+++ examples/microphone/src/prelude.rs
@@ -1,5 +1,4 @@
 pub use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
 pub use cpal::{Sample, SampleRate, Stream as CpalStream};
-pub use std::sync::atomic::{AtomicBool, Ordering};
 pub use std::sync::{Arc, Mutex};
 pub use tokio::sync::mpsc;